# backend/docs/postgres-reference/openapi.design.yaml
openapi: 3.0.3
info:
  title: School Admission Management API
  version: 1.0.0
  description: |
    REST API for enquiries, applications, workflow transitions, assessments,
    payments and reporting.

    **Conventions**
    - Base path `/api/v1`. JSON only. Timestamps are ISO-8601 UTC.
    - Auth: OAuth2 / OIDC bearer JWT. Guardians sign in with OTP; the JWT carries a `roles` claim.
    - Errors use RFC 9457 `application/problem+json`.
    - Mutating calls accept an `Idempotency-Key` header; payment calls require it.
    - Applications use optimistic locking: send `If-Match: <version>` on writes; a stale value returns `409`.
    - Lists are cursor-paginated: `?limit=25&cursor=...`.
servers:
  - url: https://api.example-school.in/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /interaction-slots:
    get:
      tags: [Scheduling]
      summary: List open slots
      parameters:
        - { name: kind, in: query, schema: { type: string, enum: [INTERACTION, CAMPUS_TOUR, ENTRANCE_TEST] } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Slots with remaining capacity
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Slot' }
  /interaction-slots/{slotId}/bookings:
    post:
      tags: [Scheduling]
      summary: Book a slot for an application
      parameters:
        - { name: slotId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [applicationId]
              properties:
                applicationId: { type: string, format: uuid }
      responses:
        '201': { description: Booked }
        '409': { description: Slot is full }

  # -------------------------------------------------------------- Assessment
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Slot:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        remaining: { type: integer }
...